Redid interrupt code to be somewhat usable.
Must register now with the IO interrupt handler, which will
place you on a list of interrupt functions to call when the
interrupt is received.
Also, implemented INT_RESCHEDULE & INT_NO_RESCHEDULE which
means the interrupt code will choose to reschedule you or
not based on what you return from the interrupt handler.