Fixed broken thread creation that existed after last checkin.
Some inline-but-outside-of-function asm() statement was being
ignored. I now consider that to be a bad idea, but it sure was
nice to use. Moved context_switch code to standalone assembly
source. A few of the threads are not getting scheduled however,
but I'm not sure why. Need to write a kernel debugger so I can
investigate a crashed system.